+++ src/Media.php
@@ -35,6 +35,7 @@ use Psr\Http\Message\StreamInterface;
  * @property string|null $variant_name
  * @property int|string|null $original_media_id
  * @property int|null $size
+ * @property string|null $alt
  * @property Carbon $created_at
  * @property Carbon $updated_at
  * @property Pivot $pivot
@@ -76,7 +77,8 @@ class Media extends Model
         'mime_type',
         'aggregate_type',
         'variant_name',
-        'original_media_id'
+        'original_media_id',
+        'alt',
     ];
 
     protected $casts = [